Demo - Status Page Widget
This is just a demonstration page for the status page widget. Deployment in a usable format would involve including an Auto or On Load refresh. The page I originally deployed this to was best suited for On Load refresh so I haven't included an auto-refresh as part of this code.